requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
boilerpy3==1.0.7
certifi==2024.2.2
charset-normalizer==3.3.2
datasets==2.19.0
dill==0.3.8
distro==1.9.0
exceptiongroup==1.2.1
filelock==3.14.0
frozenlist==1.4.1
fsspec==2024.3.1
gdown==5.2.0
h11==0.14.0
haystack-ai==2.0.1
haystack-bm25==1.0.2
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.0
idna==3.7
install==1.3.5
Jinja2==3.1.3
joblib==1.4.2
lazy-imports==0.3.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdit-plain==1.0.1
mdurl==0.1.2
monotonic==1.6
more-itertools==10.2.0
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
networkx==3.1
numpy==1.24.4
openai==1.25.1
packaging==24.0
pandas==2.0.3
pillow==10.3.0
posthog==3.5.0
pyarrow==16.0.0
pyarrow-hotfix==0.6
pydantic==2.7.1
pydantic-core==2.18.2
pypdf==4.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
regex==2024.4.28
requests==2.31.0
safetensors==0.4.3
scikit-learn==1.3.2
scipy==1.10.1
sentence-transformers==2.7.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
structlog==24.1.0
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.2.2
tqdm==4.66.4
transformers==4.40.1
typing-extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
xxhash==3.4.1
yarl==1.9.4
